# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Ve mi Tienda",
        "description": "Se inicia sesión, Auth->login, se toma el token y se ingresa arriba en el botón Authorize",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://vemitiendabackend.tests/api/v1/"
        }
    ],
    "paths": {
...
        "/company-user": {
            "get": {
                "tags": [
                    "Company"
                ],
                "summary": "Mostrar los datos de la Empresa del Usuario de la App",
                "operationId": "c81c57a22b280dd0807ec5923b9d3c7f",
                "responses": {
                    "200": {
                        "description": "Exitoso"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Company"
                ],
                "summary": "Si la Empresa no existe, la crea y si la empresa existe entonces la actualiza",
                "operationId": "73c33f62482b88b25b1a8cc11ce55c27",
                "requestBody": {
                    "description": "Datos de la Empresa",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "name",
                                    "slogan",
                                    "phone"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "sistelconet@gmail.com"
                                    },
                                    "name": {
                                        "type": "string",
                                        "format": "name",
                                        "example": "Sistelconet El Tigre"
                                    },
                                    "slogan": {
                                        "type": "string",
                                        "format": "slogan",
                                        "example": "Tu Solución en Sistemas"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "format": "phone",
                                        "example": "+584248807465"
                                    },
                                    "theme_id": {
                                        "type": "integer",
                                        "format": "phone",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Exitoso"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/company/{id}": {
            "delete": {
                "tags": [
                    "Company"
                ],
                "summary": "Borrar categoría de un Usuario App",
                "operationId": "d70242fba1f19b199fc885a114092bee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exitoso"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/user-information": {
            "get": {
                "summary": "Mostrar Información del Usuario de la App",
                "operationId": "7ba535b69ac822f82129f3610b6ac46d",
                "responses": {
                    "200": {
                        "description": "Exitoso"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "User Login",
                "description": "Login User Here",
                "operationId": "authLogin",
                "requestBody": {
                    "description": "Datos de la Empresa",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "administrador@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Login Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Login Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Desloguear usuario",
                "operationId": "7c822f4eec9424890f34674874012768",
                "responses": {
                    "200": {
                        "description": "Exitoso"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/cancel-account": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Dar de baja al usuario",
                "operationId": "73acaae339ec1746a5c1c6214654850e",
                "responses": {
                    "200": {
                        "description": "Exitoso"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
...
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Ingresa el Token con la palabra Bearer, así: Bearer <token>",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}